[macros] Add nightly toolchain and use arbitray_self_types feature as experiment
Currently this feature is missing support for Mutex or RwLock wrapped
inside an Arc which makes the proc macro to generate "manager" of
"thing" still very relevant. This feature might make things easier in
the future, but that would still not invalidate the need to have this
macro (this feature **might** just help make the implementation of that
macro simpler).